db: build out profile actions
